Improve module hierarchy
